Beta Build 3
When changing log name it will not rename the save name until a default
log exist
Make rich text on log detected UI
Remove “Time as been logged as “ when inserting time.
When track is pressed set curser to end.
When you cancel from the load dialog, it still opens the program.